kernel: [DANGER] Create sci.c and facing compilation issues again.

+++ landos/kernel/1pump/arch/x86_64/x64.c
@@ -191,9 +191,8 @@ tss_init (
     memset ( tss, 0, sizeof *tss );
 
     //ring 0 stack
-    tss->rsp0 = stack_address;  // va ?? 
-    
-    
+    tss->rsp0 = (unsigned long) stack_address;  // va ?? 
+
     //#debug
     //printf ("Stack %x\n", stack_address);
     //refresh_screen();
